# perf: use a hashmap when possible for file extension matching