@5225225 2025-04-10
Uh, that unit-tests-openbsd failure ( https://github.com/eza-community/eza/actions/runs/14377641116/job/40313754679?pr=1421 )  looks flaky. I have no clue what would have caused that, I don't *think* it's related to my changes, seeing as it can't seem to find cargo?